Patch from Ralf Corsepius <corsepiu@faw.uni-ulm.de>:
    Yet some more modifications, I would recommend to be considered before
    releasing a snapshot:

    1. Cleanup to aclocal/
    cvs rm -f aclocal/cygwin.m4
    cvs rm -f aclocal/exeext.m4

    They are neither used nor needed anymore, however they also don't
    disturb (we use autoconf-2.13's AC_EXEEXT instead, now)

    ----------

    2. rtems-rc-19990328-0.diff
    Some (minor) bug-fixes:
    * make/Templates/Makefile.inc.in: use the new installation directory
    ($(prefix)/ instead of $(prefix)/rtems/)
    * c/src/exec/score/tools/generic/Makefile.am: added line to include local.am
    * c/src/exec/score/tools/*/configure.in: added CVS Id header

    ----------

    3. rtems-rc-19990328-1.diff
    Enhancements and cleanups to autogen, rtems-polish.sh, configure.in etc.

    * autogen: Use the file "VERSION" to detect RTEMS toplevel directory,
    extended usage-message, use "find -print"
    * c/update-tools/cipolish: New script to beautify configure.in scripts
    * c/update-tools/rtems-polish.sh: Use the file "VERSION" to detect RTEMS
    toplevel directory, extended usage-message, added variable for perl
    scripts' subdirectory, use "find -print", cipolish support, new options
    -ac -am -ci.
    * aclocal/*.m4, configure.in: moved some AC_SUBST lines to aclocal/*.m4
    (reduces size of configure.in
    scripts, eases splitting configure.in scripts).

    ----------

+++ c/update-tools/rtems-polish.sh
@@ -1,8 +1,70 @@
 #!/bin/sh
 
+# $Id$
+
+#
+# Search RTEMS source tree for autoconf Makefile.ins and automake
+# Makefile.ams and run c/update-tools/acpolish rsp. c/update-tool/ampolish
+# on them.
+#
+# To be run from the toplevel directory of the source-tree
+#
+
+progname=`basename $0`
+
+perltools=c/update-tools
+ac_do=""
+am_do=""
+ci_do=""
+
+usage()
+{
+  echo
+  echo "usage: ./${perltools}/${progname} [-h][-ac|-am|-ci]";
+  echo
+  echo "options:"
+  echo "	-h .. display this message and exit";
+  echo "	-ac .. run acpolish on all autoconf Makefile.ins"
+  echo "	-am .. run ampolish on all automake Makefile.ams"
+  echo "	-ci .. run cipolish on all configure.in scripts"
+  echo
+  exit 1;
+}
+
+if test ! -f VERSION; then
+  echo "${progname}:"
+  echo "        Please change directory to RTEMS's toplevel directory"
+  exit 1;
+fi
+
+while test $# -gt 0; do
+case $1 in
+-h|--he|--hel|--help)
+  usage ;;
+-ac)
+  ac_do="yes";
+  shift ;;
+-am)
+  am_do="yes";
+  shift ;;
+-ci)
+  ci_do="yes";
+  shift ;;
+-*) echo "unknown option $1" ;
+  usage ;;
+*) echo "invalid parameter $1" ;
+  usage ;;
+esac
+done
+
+if test -z "$ac_do" && test -z "$am_do" && test -z "$ci_do"; then
+  usage
+fi
+
 pwd=`pwd`;
 
-ac_files=`find . -name Makefile.in`;
+if test -n "$ac_do"; then
+ac_files=`find . -name 'Makefile.in' -print`;
 for f in $ac_files; do
   i=`dirname $f`
   dest="$i"
@@ -10,21 +72,37 @@ for f in $ac_files; do
     echo "polishing : $dest/Makefile.in"
     ( cd $dest; 
       mv Makefile.in Makefile.in~;
-      $pwd/c/update-tools/acpolish <Makefile.in~ >Makefile.in
+      ${pwd}/${perltools}/acpolish <Makefile.in~ >Makefile.in
       rm Makefile.in~
     )
   fi
 done
+fi
 
-am_files=`find . -name Makefile.am`;
+if test -n "$am_do"; then
+am_files=`find . -name 'Makefile.am' -print`;
 for f in $am_files; do
   i=`dirname $f`
   dest="$i"
     echo "polishing : $dest/Makefile.am"
     ( cd $dest; 
       mv Makefile.am Makefile.am~;
-      $pwd/c/update-tools/ampolish <Makefile.am~ >Makefile.am
+      ${pwd}/${perltools}/ampolish <Makefile.am~ >Makefile.am
       rm Makefile.am~
     )
 done
+fi
 
+if test -n "$ci_do"; then
+ci_files=`find . -name 'configure.in' -print`;
+for f in $ci_files; do
+  i=`dirname $f`
+  dest="$i"
+    echo "polishing : $dest/configure.in"
+    ( cd $dest; 
+      mv configure.in configure.in~;
+      ${pwd}/${perltools}/cipolish <configure.in~ >configure.in
+      rm configure.in~
+    )
+done
+fi
